+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
-- * b881d12 - 2018-11-17: Pre-Release v9.4.36.dev3 
+- * 5d32bc5 - 2018-11-17: Fixed issues with grabbing redirect url for session POST method. Fixed login issue for FileList. Fixed issue with parsing torrents from Zooqle and YGGTorrent. 
+- * 5acf339 - 2018-11-17: Pre-Release v9.4.36.dev3 
 - * 34dd950 - 2018-11-17: Fixed issue #322 - NoneType error when trying to view Trakt &#x60;most anticipated&#x60; shows due to indexer failing to have artwork available. Refactored error page to new UI look. 
 - * 3299a1e - 2018-11-17: Pre-Release v9.4.36.dev2 
 - * 9048cd4 - 2018-11-17: Refactored DanishBits torrent provider.